+++ t/check_arcconf.t
@@ -6,7 +6,7 @@ BEGIN {
 
 use strict;
 use warnings;
-use Test::More tests => 22;
+use Test::More tests => 27;
 use test;
 
 # NOTE: this plugin has side effect of changing dir
@@ -15,73 +15,54 @@ ok(chdir(TESTDIR), "dir changed");
 my $cwd = Cwd::cwd();
 ok(defined($cwd), "cwd set");
 
-if (1) {
-my $plugin = arcconf->new(
-	commands => {
-		'getstatus' => ['<', $cwd. '/data/arcconf/1/getstatus'],
-		'getconfig' => ['<', $cwd. '/data/arcconf/1/getconfig'],
+my @tests = (
+	{
+		status => OK,
+		getstatus => '1/getstatus',
+		getconfig => '1/getconfig',
+		message => 'Controller:Optimal, Logical Device 0:Optimal, Drives: 3LQ1WEH4,3LQ1WEWL=Online',
+	},
+	{
+		status => OK,
+		getstatus => '1/getstatus',
+		getconfig => '2/getconfig',
+		message => 'Controller:Optimal, Battery Status: Optimal, Battery Capacity Remaining: 100%, Battery Time: 3d16h0m, Logical Device 0:Optimal, Drives: 5000C5000B426E88,500000E01BEFF141=Online',
+	},
+	{
+		status => OK,
+		getstatus => '3/getstatus',
+		getconfig => '3/getconfig',
+		message => 'Controller:Optimal, Logical device #0: Build/Verify: In Progress 11%, ZMM Status: ZMM Optimal, Logical Device 0(Volume01):Optimal, Drives: *******,*******,*******,*******=Online',
+	},
+	{
+		status => OK,
+		getstatus => '4/arcconf_getstatus_1.out',
+		getconfig => '4/arcconf_getconfig_1_al.out',
+		message => 'Controller:Okay, Logical Device 1(MailMerak):Okay, Drives: 9QJ2C6NX,GTA060PBGHX61F=Online',
+	},
+	{
+		status => OK,
+		getstatus => 'issue31/getstatus',
+		getconfig => 'issue31/getconfig',
+		message => 'Controller:Optimal, Logical Device 0(OS):Optimal, Logical Device 1(SSDSTORAGEPOOL):Optimal, Logical Device 2(HDDSTORAGEPOOL):Optimal, Drives: xxxx,xxxx,xxxx,xxxx,xxxx,xxxx,xxxx,xxxx,xxxx,xxxx,xxxx=Online',
 	},
 );
 
+foreach my $test (@tests) {
+	my $plugin = arcconf->new(
+		commands => {
+			getstatus => ['<', $cwd . '/data/arcconf/' . $test->{getstatus}],
+			getconfig => ['<', $cwd . '/data/arcconf/' . $test->{getconfig}],
+		},
+	);
 
-ok($plugin, "plugin created");
-$plugin->check;
-ok(1, "check ran");
-ok(defined($plugin->status), "status code set");
-ok($plugin->status == OK, "status OK");
-print "[".$plugin->message."]\n";
-ok($plugin->message eq 'Controller:Optimal, Logical Device 0:Optimal');
-}
+	ok($plugin, "plugin created");
 
-if (1) {
-my $plugin = arcconf->new(
-	commands => {
-		'getstatus' => ['<', $cwd. '/data/arcconf/1/getstatus'],
-		'getconfig' => ['<', $cwd. '/data/arcconf/2/getconfig'],
-	},
-);
+	$plugin->check;
+	ok(1, "check ran");
 
-ok($plugin, "plugin created");
-$plugin->check;
-ok(1, "check ran");
-ok(defined($plugin->status), "status code set");
-ok($plugin->status == OK, "status OK");
-print "[".$plugin->message."]\n";
-ok($plugin->message eq 'Controller:Optimal, Battery Status: Optimal, Battery Capacity Remaining: 100%, Battery Time: 3d16h0m, Logical Device 0:Optimal');
-}
-
-if (1) {
-my $plugin = arcconf->new(
-	commands => {
-		'getstatus' => ['<', $cwd. '/data/arcconf/3/getstatus'],
-		'getconfig' => ['<', $cwd. '/data/arcconf/3/getconfig'],
-	},
-);
-
-
-ok($plugin, "plugin created");
-$plugin->check;
-ok(1, "check ran");
-ok(defined($plugin->status), "status code set");
-ok($plugin->status == OK, "status OK");
-print "[".$plugin->message."]\n";
-ok($plugin->message eq 'Controller:Optimal, Logical device #0: Build/Verify: In Progress 11%, ZMM Status: ZMM Optimal, Logical Device 0(Volume01):Optimal');
-}
-
-if (1) {
-my $plugin = arcconf->new(
-	commands => {
-		'getstatus' => ['<', $cwd. '/data/arcconf/4/arcconf_getstatus_1.out'],
-		'getconfig' => ['<', $cwd. '/data/arcconf/4/arcconf_getconfig_1_al.out'],
-	},
-);
-
-
-ok($plugin, "plugin created");
-$plugin->check;
-ok(1, "check ran");
-ok(defined($plugin->status), "status code set");
-ok($plugin->status == OK, "status OK");
-print "[".$plugin->message."]\n";
-ok($plugin->message eq 'Controller:Okay, Logical Device 1(MailMerak):Okay');
+	ok(defined($plugin->status), "status code set");
+	ok($plugin->status == $test->{status}, "status code (got:".$plugin->status." exp:".$test->{status}.")");
+	print "[".$plugin->message."]\n";
+	ok($plugin->message eq $test->{message}, "status message");
 }